1.线工单日志中新增不良数量字段 2.设备工单日志中的不良数量字段类型修改为浮点型 3.新增物料请求 及 物料请求结束

+++ mpm_web_api/model/m_wo/wo_machine_cur_log.cs
@@ -28,7 +28,7 @@ namespace mpm_web_api.model.m_wo
         /// <summary>
         /// 不良数量
         /// </summary>
-        public int bad_quantity { get; set; }
+        public decimal bad_quantity { get; set; }
         /// <summary>
         /// 达成率
         /// </summary>
